+++ app/build.gradle
@@ -36,6 +36,7 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
         useIR = true
+        freeCompilerArgs = ['-Xinline-classes']
     }
     buildFeatures {
         compose true
