+++ app/src/main/java/com/louis993546/composecomposer/ui/renderer/ColumnNode.kt
@@ -3,7 +3,7 @@ package com.louis993546.composecomposer.ui.renderer
 import androidx.compose.foundation.layout.Column
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.louis993546.composecomposer.Node
+import com.louis993546.composecomposer.data.model.Node
 import com.louis993546.composecomposer.util.PreviewTheme
 
 @Composable
