+++ app/src/main/java/com/louis993546/composecomposer/ui/renderer/PageRenderer.kt
@@ -5,13 +5,12 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.unit.dp
-import com.louis993546.composecomposer.Page
+import com.louis993546.composecomposer.data.model.Page
 
 @Composable
 fun PageRenderer(
