+++ app/src/main/java/com/louis993546/composecomposer/util/Util.kt
@@ -1,12 +1,12 @@
 package com.louis993546.composecomposer.util
 
 import androidx.compose.runtime.Composable
-import com.louis993546.composecomposer.Id
+import com.louis993546.composecomposer.data.model.Id
 import com.louis993546.composecomposer.ui.theme.ComposeComposerTheme
 import kotlin.random.Random
 
 // TODO it'd be nice i can make sure it never collide
-fun randId(): Id = Random.nextInt()
+fun randId(): Id = Id(Random.nextInt())
 
 val <T> T.exhaustive: T
     get() = this
